+++ projects/client/src/routes/shows/recommended/+page.svelte
@@ -2,7 +2,7 @@
   import TraktPage from "$lib/components/layout/TraktPage.svelte";
   import * as m from "$lib/features/i18n/messages";
 
-  import RecommendedPaginatedList from "$lib/sections/lists/RecommendedPaginatedList.svelte";
+  import RecommendedPaginatedList from "$lib/sections/lists/recommended/RecommendedPaginatedList.svelte";
   import { DEFAULT_SHARE_SHOW_COVER } from "$lib/utils/constants";
 </script>
